Rename bucket move job
