+++ searchcore/src/vespa/searchcore/proton/server/bucketmovejob.cpp
@@ -1,6 +1,6 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
-#include "bucketmovejobv2.h"
+#include "bucketmovejob.h"
 #include "imaintenancejobrunner.h"
 #include "ibucketstatechangednotifier.h"
 #include "iclusterstatechangednotifier.h"
