+++ searchcore/src/vespa/searchcore/proton/server/maintenance_jobs_injector.cpp
@@ -1,6 +1,6 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
-#include "bucketmovejobv2.h"
+#include "bucketmovejob.h"
 #include "heart_beat_job.h"
 #include "job_tracked_maintenance_job.h"
 #include "lid_space_compaction_job.h"
